requirements.txt
django
djangorestframework
opencv-python
pytesseract
pillow
django-cors-headers